Extract methods for finding categories
