+++ src/flickypedia/apis/wikimedia/api.py
@@ -8,7 +8,6 @@ designed to be used with any auth approach.
 See https://api.wikimedia.org/wiki/Authentication
 """
 
-import re
 import typing
 from xml.etree import ElementTree as ET
 
@@ -23,11 +22,14 @@ from .exceptions import (
 from .base import HttpxImplementation
 from .languages import LanguageMatch, order_language_list
 
+from .category_methods import CategoryMethods
 from .structured_data_methods import StructuredDataMethods
 from .validator_methods import ValidatorMethods
 
 
-class WikimediaApi(HttpxImplementation, StructuredDataMethods, ValidatorMethods):
+class WikimediaApi(
+    HttpxImplementation, CategoryMethods, StructuredDataMethods, ValidatorMethods
+):
     def _get(self, params: dict[str, str]) -> typing.Any:
         return self._get_json(params=params)
         return self._request(method="GET", params={**params, "format": "json"})
@@ -123,76 +125,6 @@ class WikimediaApi(HttpxImplementation, StructuredDataMethods, ValidatorMethods)
             raise RuntimeError(f"Unexpected result from upload API: {upload_resp!r}")
 
         return upload_resp["upload"]["filename"]  # type: ignore
-
-    def find_matching_categories(self, query: str) -> list[str]:
-        """
-        Return a list of categories that might match this query.
-
-        This can be used to build an autocomplete interface for categories,
-        e.g. if the user types "a" we can suggest categories that
-        include the letter "a":
-
-            >>> find_matching_categories(query='a')
-            ['A',
-             'Aircraft of Uzbekistan',
-             'Aircraft of Malawi',
-             'Aircraft in Portugal',
-             'Aircraft in Sierra Leone',
-             'Aircraft in Malawi',
-             'Architectural elements in Russia',
-             'Aircraft in Thailand',
-             'Airliners in Russian service',
-             'Airliners in Malawian service']
-
-        Note: the results from this API may vary over time, or even
-        different requests. For example, there is a test in test_wikimedia.py
-        that makes the same query as the example, but returns a slightly
-        different set of categories.
-        """
-        # I found this API action by observing the network traffic in
-        # the Upload Wizard.
-        #
-        # I changed the response format from JSON to XML to give better
-        # visibility into the structure of the response -- the JSON
-        # gives an opaque list of lists.
-        #
-        # See https://www.mediawiki.org/wiki/API:Opensearch
-        resp = self.client.request(
-            "GET",
-            url="https://commons.wikimedia.org/w/api.php",
-            params={
-                "action": "opensearch",
-                "format": "xml",
-                "limit": "10",
-                "search": query,
-                # Here "14" is the namespace for categories; see
-                # https://commons.wikimedia.org/wiki/Help:Namespaces
-                "namespace": "14",
-            },
-        )
-
-        xml = ET.fromstring(resp.text)
-
-        # The XML response is of the form:
-        #
-        #   <?xml version="1.0"?>
-        #   <SearchSuggestion xmlns="http://opensearch.org/searchsuggest2" version="2.0">
-        #     <Query xml:space="preserve">a</Query>
-        #     <Section>
-        #       <Item>
-        #         <Text xml:space="preserve">Category:A</Text>
-        #         <Url xml:space="preserve">https://commons.wikimedia.org/wiki/Category:A</Url>
-        #         <Image source="https://upload.wikimedia.org/wikipedia/commons/thumb/8/85/ICS_Alfa.svg/50px-ICS_Alfa.svg.png" width="50" height="50"/>
-        #       </Item>
-        #       …
-        #
-        # We're interested in those <Text> values.
-        namespaces = {"": "http://opensearch.org/searchsuggest2"}
-
-        return [
-            re.sub(r"^Category:", "", text_elem.text)  # type: ignore
-            for text_elem in xml.findall(".//Text", namespaces=namespaces)
-        ]
 
     def find_matching_languages(self, query: str) -> list[LanguageMatch]:
         """
